requirements.txt
showdownpy
requests
pysimplegui
django